fixed post and put routes for tags and products

+++ Develop/routes/api/product-routes.js
@@ -5,11 +5,20 @@ const { Product, Category, Tag, ProductTag } = require('../../models');
 
 // get all products
 router.get('/', (req, res) => {
+  console.log('line 8, testing to fail');
   // find all products
-  Product.findAll()
+  Product.findAll({
+    include: [
+      Category,
+      {
+        model: Tag,
+        through: ProductTag
+      }
+    ]
+  })
     .then(dbProductData => res.json(dbProductData))
     .catch(err => {
-      console.log(err);
+      console.log('line 13', err);
       res.status(500).json(err);
     });
   // be sure to include its associated Category and Tag data
@@ -21,7 +30,25 @@ router.get('/:id', (req, res) => {
   Product.findOne({
     where: {
       id: req.params.id
-    }
+    },
+    attributes: [
+      'id',
+      'product_name',
+      'price',
+      'stock',
+      'category_id'
+    ],
+    // be sure to include its associated Category and Tag data
+    include: [
+      {
+        model: Category,
+        attributes: ['category_name']
+      },
+      {
+        model: Tag,
+        attributes: ['tag_name']
+      }
+    ]
   })
     .then(dbProductData => {
       if (!dbProductData) {
@@ -34,7 +61,6 @@ router.get('/:id', (req, res) => {
       console.log(err);
       res.status(500).json(err);
     });
-  // be sure to include its associated Category and Tag data
 });
 
 // create new product
@@ -71,44 +97,51 @@ router.post('/', (req, res) => {
 
 // update product
 router.put('/:id', (req, res) => {
+  const bodyCopy = {...req.body };
+  delete bodyCopy.tagIds;
+  console.log('------------------')
+  console.log(req.body);
+  console.log(bodyCopy);
   // update product data
-  Product.update(req.body, {
+  Product.update(bodyCopy, {
     where: {
       id: req.params.id,
     },
   })
-    .then((product) => {
-      // find all associated tags from ProductTag
-      return ProductTag.findAll({ where: { product_id: req.params.id } });
-    })
-    .then((productTags) => {
+    .then(async(product) => {
+      if (req.body.tagIds && req.body.tagIds.length) {
+              // find all associated tags from ProductTag
+      const productTags = await ProductTag.findAll({ where: { product_id: req.params.id } });
       // get list of current tag_ids
       const productTagIds = productTags.map(({ tag_id }) => tag_id);
       // create filtered list of new tag_ids
       const newProductTags = req.body.tagIds
-        .filter((tag_id) => !productTagIds.includes(tag_id))
-        .map((tag_id) => {
-          return {
+      .filter((tag_id) => !productTagIds.includes(tag_id))
+      .map((tag_id) => {
+        return {
             product_id: req.params.id,
             tag_id,
           };
         });
-      // figure out which ones to remove
-      const productTagsToRemove = productTags
+        // figure out which ones to remove
+        const productTagsToRemove = productTags
         .filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
         .map(({ id }) => id);
-
-      // run both actions
-      return Promise.all([
-        ProductTag.destroy({ where: { id: productTagsToRemove } }),
-        ProductTag.bulkCreate(newProductTags),
-      ]);
-    })
-    .then((updatedProductTags) => res.json(updatedProductTags))
-    .catch((err) => {
-      // console.log(err);
-      res.status(400).json(err);
-    });
+        console.log('line 127');
+        
+        // run both actions
+        return Promise.all([
+          ProductTag.destroy({ where: { id: productTagsToRemove } }),
+          ProductTag.bulkCreate(newProductTags),
+        ])
+      }
+      res.json(product);
+      })
+      .then(dbProductData => res.json(dbProductData))
+      .catch((err) => {
+        console.log(err);
+        res.status(400).json(err);
+      });
 });
 
 router.delete('/:id', (req, res) => {
@@ -123,7 +156,7 @@ router.delete('/:id', (req, res) => {
       res.status(404).json({ message: 'No product found with this id' });
       return;
     }
-    res.json(dbCategoryData);
+    res.json(dbProductData);
   })
   .catch(err => {
     console.log(err);
